server/requirements.txt
django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
psycopg2-binary==2.9.10